Refine resume page
