--- src/resume/Content/Qualification/Education/index.js
@@ -1,2 +1,0 @@
-import Education from "./Education";
-export default Education;
