--- src/resume/Content/Qualification/Personal/index.js
@@ -1,2 +1,0 @@
-import Personal from "./Personal";
-export default Personal;
