--- src/resume/Content/Qualification/Skills/index.js
@@ -1,2 +1,0 @@
-import Skills from "./Skills";
-export default Skills;
