--- src/resume/Content/Qualification/WorkExp/index.js
@@ -1,2 +1,0 @@
-import WorkExp from "./WorkExp";
-export default WorkExp;
