--- src/resume/Content/Qualification/index.js
@@ -1,2 +1,0 @@
-import Qualification from "./Qualification";
-export default Qualification;
