--- src/resume/Content/index.js
@@ -1,2 +1,0 @@
-import Content from "./Content";
-export default Content;
