+++ src/resume/Qualification/Skills/Styles.js
