+++ src/resume/Qualification/WorkExp/index.js
@@ -1,6 +1,6 @@
 import uniqid from 'uniqid';
-import TextAligned from '../../../../styles/TextAligned';
-import { WorkExpUnitDescWrapper, WorkExpUnitHeaderWrapper, WorkExpUnitSubWrapper, WorkExpWrapper, WorkExpUnitWrapper } from './styles';
+import TextAligned from '../../../styles/TextAligned';
+import { WorkExpUnitDescWrapper, WorkExpUnitHeaderWrapper, WorkExpUnitSubWrapper, WorkExpWrapper, WorkExpUnitWrapper } from './Styles';
 
 const WorkExp = ({workExp}) => {
     return (
@@ -18,16 +18,16 @@ const WorkExpUnit = ({workExpUnit}) => {
     return (
         <WorkExpUnitWrapper>
             <WorkExpUnitHeaderWrapper>
-                <TextAligned weight="700">{workExpUnit.jobTitle}</TextAligned>
-                {workExpUnit.workFrom && workExpUnit.workTo ? <TextAligned aligned="right">{workExpUnit.workFrom} - {workExpUnit.workTo}</TextAligned> : <></>}
+                <TextAligned weight="700" contentEditable="true">{workExpUnit.jobTitle}</TextAligned>
+                {workExpUnit.workFrom && workExpUnit.workTo ? <TextAligned aligned="right" contentEditable="true">{workExpUnit.workFrom} - {workExpUnit.workTo}</TextAligned> : <></>}
             </WorkExpUnitHeaderWrapper>
             <WorkExpUnitSubWrapper>
-                <TextAligned>{workExpUnit.orgName}</TextAligned>
-                <TextAligned aligned="right">{workExpUnit.orgPlace}</TextAligned>
+                <TextAligned contentEditable="true">{workExpUnit.orgName}</TextAligned>
+                <TextAligned aligned="right" contentEditable="true">{workExpUnit.orgPlace}</TextAligned>
             </WorkExpUnitSubWrapper>
             <WorkExpUnitDescWrapper>
                     {workExpUnit.achievements.split('. ').map((sentence) => {
-                        return <div key={uniqid()} className="f-size-80">{sentence}</div>
+                        return <div key={uniqid()} className="f-size-80" contentEditable="true">{sentence}</div>
                     })}
             </WorkExpUnitDescWrapper>
