--- src/resume/Resume.js
@@ -1,9 +1,0 @@
-import Content from "./Content";
-
-const Resume = ({collection, font}) => {
-    return (
-        <Content collection={collection} font={font}/>
-    )
-}
-
-export default Resume;
